+++ CHANGELOG.rst
@@ -21,11 +21,15 @@ Added
 - Added `pyxem.utils.calibration_utils.Calibration` class  for calibrating the signal axes of a 4-D STEM dataset(#993)
 - Added `pyxem.signals.Diffraction2D.get_azimuthal_integral1D` method to calculate the azimuthal integral of a 2D diffraction pattern (#1008)
 - Added example for doing azimuthal integration of a 2d diffraction pattern (#1009)
+- Added `pyxem.signals.CommonDiffraction.get_virtual_image` method to calculate multiple virtual images
+  from a 4D STEM dataset (#1014)
 
 Deprecated
 ----------
 - The module & all functions within `utils.reduced_intensity1d` are deprecated in favour of using the methods of `ReducedIntensity1D` (#994).
 - Deprecated `CalibrationGenerator` in favour of `pyxem.utils.calibration_utils.Calibration` class (#1000)
+- Deprecated `pyxem.generators.virtual_image_generator.VirtualImageGenerator` in
+  favor of `pyxem.signals.CommonDiffraction.get_virtual_image` (#1014)
 
 Deleted
 -------